# Remove an explicitly-set-aside scratch register per class.