--- a/src/ion/data_structures.rs
+++ b/src/ion/data_structures.rs
@@ -351,7 +351,8 @@ pub struct Env<'a, F: Function> {
     pub spillslots: Vec<SpillSlotData>,
     pub slots_by_size: Vec<SpillSlotList>,
 
-    pub extra_spillslot: Vec<Option<Allocation>>,
+    pub extra_spillslots_by_class: [SmallVec<[Allocation; 2]>; 2],
+    pub preferred_victim_by_class: [PReg; 2],
 
     // Program moves: these are moves in the provided program that we
     // handle with our internal machinery, in order to avoid the