--- a/src/ion/liveranges.rs
+++ b/src/ion/liveranges.rs
@@ -109,6 +109,13 @@ impl<'a, F: Function> Env<'a, F> {
         for &preg in &self.env.fixed_stack_slots {
             self.pregs[preg.index()].is_stack = true;
         }
+        for class in 0..self.preferred_victim_by_class.len() {
+            self.preferred_victim_by_class[class] = self.env.non_preferred_regs_by_class[class]
+                .last()
+                .or(self.env.preferred_regs_by_class[class].last())
+                .cloned()
+                .unwrap_or(PReg::invalid());
+        }
         // Create VRegs from the vreg count.
         for idx in 0..self.func.num_vregs() {
             // We'll fill in the real details when we see the def.