--- a/src/lib.rs
+++ b/src/lib.rs
@@ -248,10 +248,13 @@ pub struct SpillSlot {
 }
 
 impl SpillSlot {
+    /// The maximum spillslot index.
+    pub const MAX: usize = (1 << 24) - 1;
+
     /// Create a new SpillSlot of a given class.
     #[inline(always)]
     pub fn new(slot: usize, class: RegClass) -> Self {
-        debug_assert!(slot < (1 << 24));
+        debug_assert!(slot <= Self::MAX);
         SpillSlot {
             bits: (slot as u32) | (class as u8 as u32) << 24,
         }
@@ -1250,25 +1253,11 @@ pub struct MachineEnv {
     /// but still better than spilling.
     pub non_preferred_regs_by_class: [Vec<PReg>; 2],
 
-    /// One scratch register per class. This is needed to perform
-    /// moves between registers when cyclic move patterns occur. The
-    /// register should not be placed in either the preferred or
-    /// non-preferred list (i.e., it is not otherwise allocatable).
-    ///
-    /// Note that the register allocator will freely use this register
-    /// between instructions, but *within* the machine code generated
-    /// by a single (regalloc-level) instruction, the client is free
-    /// to use the scratch register. E.g., if one "instruction" causes
-    /// the emission of two machine-code instructions, this lowering
-    /// can use the scratch register between them.
-    pub scratch_by_class: [PReg; 2],
-
     /// Some `PReg`s can be designated as locations on the stack rather than
     /// actual registers. These can be used to tell the register allocator about
     /// pre-defined stack slots used for function arguments and return values.
     ///
-    /// `PReg`s in this list cannot be used as a scratch register or as an
-    /// allocatable regsiter.
+    /// `PReg`s in this list cannot be used as an allocatable register.
     pub fixed_stack_slots: Vec<PReg>,
 }