@cfallin 2022-05-21
Benchmark results with Cranelift: `blake3-scalar` gets 5-8% faster, runtimes otherwise unaffected.

```
execution :: cycles :: benchmarks-next/blake3-scalar/benchmark.wasm

  Δ = 26130.32 ± 5035.33 (confidence = 99%)

  new.so is 1.05x to 1.08x faster than old.so!
  old.so is 0.93x to 0.95x faster than new.so!

  [381216 389967.40 419786] new.so
  [407474 416097.72 445626] old.so

execution :: nanoseconds :: benchmarks-next/blake3-scalar/benchmark.wasm

  Δ = 6874.52 ± 1325.00 (confidence = 99%)

  new.so is 1.05x to 1.08x faster than old.so!
  old.so is 0.93x to 0.95x faster than new.so!

  [100317 102615.16 110455] new.so
  [107217 109489.68 117256] old.so
```